Split PageStats and ResultPagers into separate files
